even more superficial refactoring
